Use string append instead of appendToFile
